requirements.txt
torch
torchvision
torchaudio      # Часто идет вместе с torch/torchvision, но лучше указать явно
matplotlib      # Для визуализации
numpy           # Основная библиотека для вычислений, часто используется неявно
scikit-learn    # Понадобится для PCA или t-SNE при визуализации латентного пространства
PyYAML          # Для чтения файла конфигурации config.yaml (если будем его использовать)